Adding changes so that we store state in a volume and create the env file from the template on first deployment

+++ docker/deploy.sh
@@ -1,4 +1,15 @@
 #!/bin/bash
+if ! [ -n "$BASH_VERSION" ];then
+    echo "this is not bash, calling self with bash....";
+    SCRIPT=$(readlink -f "$0")
+    /bin/bash $SCRIPT
+    exit;
+fi
+
+DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
+
+cd $HOME
+mkdir -p polr-data/storage
 
 # Remove any existing polr container
 docker kill polr-container
@@ -8,4 +19,5 @@ docker rm polr-container
 docker run -d \
 -p 80:80 \
 --name polr-container \
+--volume $HOME/polr-data:/data \
 polr
